Revert "Squashed commit of the following:"
This reverts commit d70292229dbdfc4a3e3b1a47e17f3d3297b703ab.

+++ src_c/scrap.c
@@ -208,7 +208,7 @@ _scrap_get_scrap(PyObject *self, PyObject *args)
     char *scrap = NULL;
     PyObject *retval;
     char *scrap_type;
-    size_t count;
+    unsigned long count;
 
     PYGAME_SCRAP_INIT_CHECK();
